requirements.txt
gradio
flask
python-dotenv
pypdf
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
gunicorn
torch
transformers
numpy